flask-web-library/requirements.txt
flask
psycopg2-binary
sqlalchemy
numpy
django
djangorestframework
pgvector
requests
django-cors-headers